End UI Work
Got the majority of the end pad UI working now, Just working on out of bounds ui next.

+++ Assets/Scripts/SaveScript.cs
@@ -27,65 +27,66 @@ public struct LevelScores
 
 public class SaveScript : MonoBehaviour
 {
-	public List<LevelScores> PerLevelData;
+	public List<LevelScores> LevelData;
 
 	private void Start()
 	{
-		//if (File.Exists(Application.persistentDataPath + "/gamedata.ini"))
-		//{
-		//	LoadData();
-		//}
-		//else
-		//{
-		//	CreateFile();
-		//}
-		PerLevelData = new List<LevelScores>();
+		LevelData = new List<LevelScores>();
 
 		for (int i = 0; i < 20; i++)
 		{
-			PerLevelData.Add(new LevelScores());
+			LevelData.Add(new LevelScores());
 		}
 
-		for (int i = 0; i < PerLevelData.Count; i++)
+		Setup();
+	}
+
+	// Initilise The File if it doesn't exsist
+	private void Setup()
+	{
+		// Goes through the Level Scores and sets up the level names to equal "Level##"
+		for (int i = 0; i < LevelData.Count; i++)
 		{
 			LevelScores Test = new LevelScores();
 			Test.LevelName = "Level" + (i + 1);
-			PerLevelData[i] = Test;
+			LevelData[i] = Test;
 		}
 
-		Debug.Log(PerLevelData.Count);
-		Debug.Log(PerLevelData[0].LevelName);
-
+		// Sets up the default values for all elements in each Level Score
 		if (!File.Exists(Application.persistentDataPath + "/gamedata.ini"))
 		{
-			for (int i = 0; i < PerLevelData.Count; i++)
+			// Goes through the LevelData and sets it up to have default values
+			for (int i = 0; i < LevelData.Count; i++)
 			{
+				// Creates a temp level scores to hold the new values
 				LevelScores Update = new LevelScores();
+
+				// Sets up the default values
 				Update.BestTime = 9999999;
 				Update.SecondBestTime = 9999999;
 				Update.ThirdBestTime = 9999999;
-				PerLevelData[i] = Update;
-				Debug.Log("Reset Scores");
+				Update.BestTimeName = LevelData[i].BestTimeName;
+				Update.SecondBestName = LevelData[i].SecondBestName;
+				Update.ThirdBestName = LevelData[i].ThirdBestName;
+				Update.StarTime = LevelData[i].StarTime;
+				Update.LastTime = LevelData[i].LastTime;
+				Update.LevelName = LevelData[i].LevelName;
 
-				Debug.Log(PerLevelData[0].BestTime);
+				// Makes the changes to the LevelData
+				LevelData[i] = Update;
 
+				// Saves the new values
 				SaveData();
 			}
 		}
 		else
 		{
+			// Loads the data if the file already exsists
 			LoadData();
 		}
 	}
 
-
-	public void CreateFile()
-	{
-		FileStream DataFile;
-		DataFile = File.Create(Application.persistentDataPath + "/gamedata.ini");
-		DataFile.Close();
-	}
-
+	// Saves the Level Scores to the file
 	public void SaveData()
 	{
 		// does stuff with binrary one assumes :)
@@ -94,43 +95,39 @@ public class SaveScript : MonoBehaviour
 		FileStream DataFile;
 
 		// creates a file in the data path /C/users/appdata/......
-		//if (!File.Exists(Application.persistentDataPath + "/gamedata.ini"))
-		//{
+		if (!File.Exists(Application.persistentDataPath + "/gamedata.ini"))
+		{
 			DataFile = File.Create(Application.persistentDataPath + "/gamedata.ini");
-		//}
-		//else
-		//{
-		//	DataFile = File.Open(Application.persistentDataPath + "/gamedata.ini", FileMode.Open);
-		//}
+		}
+		else
+		{
+			DataFile = File.Open(Application.persistentDataPath + "/gamedata.ini", FileMode.Open);
+		}
 
 		// creates an instance of the game data class...
 		GameData Data = new GameData();
 
 		// populating the new instance with the current values in the game
-		Data.Level1 = PerLevelData[0];
-		Data.Level2 = PerLevelData[1];
-		Data.Level3 = PerLevelData[2];
-		Data.Level4 = PerLevelData[3];
-		Data.Level5 = PerLevelData[4];
-		Data.Level6 = PerLevelData[5];
-		Data.Level7 = PerLevelData[6];
-		Data.Level8 = PerLevelData[7];
-		Data.Level9 = PerLevelData[8];
-		Data.Level10 = PerLevelData[9];
-		Data.Level11 = PerLevelData[10];
-		Data.Level12 = PerLevelData[11];
-		Data.Level13 = PerLevelData[12];
-		Data.Level14 = PerLevelData[13];
-		Data.Level15 = PerLevelData[14];
-		Data.Level16 = PerLevelData[15];
-		Data.Level17 = PerLevelData[16];
-		Data.Level18 = PerLevelData[17];
-		Data.Level19 = PerLevelData[18];
-		Data.Level20 = PerLevelData[19];
-
-		Debug.Log(PerLevelData[0].LastTime);
-		Debug.Log(PerLevelData[0].BestTime);
-		Debug.Log(PerLevelData[0].SecondBestTime);
+		Data.Level1 = LevelData[0];
+		Data.Level2 = LevelData[1];
+		Data.Level3 = LevelData[2];
+		Data.Level4 = LevelData[3];
+		Data.Level5 = LevelData[4];
+		Data.Level6 = LevelData[5];
+		Data.Level7 = LevelData[6];
+		Data.Level8 = LevelData[7];
+		Data.Level9 = LevelData[8];
+		Data.Level10 = LevelData[9];
+		Data.Level11 = LevelData[10];
+		Data.Level12 = LevelData[11];
+		Data.Level13 = LevelData[12];
+		Data.Level14 = LevelData[13];
+		Data.Level15 = LevelData[14];
+		Data.Level16 = LevelData[15];
+		Data.Level17 = LevelData[16];
+		Data.Level18 = LevelData[17];
+		Data.Level19 = LevelData[18];
+		Data.Level20 = LevelData[19];
 
 		// Converts to binrary, using the data from the data thingy in a data file
 		BinFormat.Serialize(DataFile, Data);
@@ -139,7 +136,7 @@ public class SaveScript : MonoBehaviour
 		DataFile.Close();
 	}
 
-
+	// Load the Level Scores from the file to the LevelData List
 	public void LoadData()
 	{
 		// checks to see if the file exsists, duh...
@@ -157,31 +154,31 @@ public class SaveScript : MonoBehaviour
 			DataFile.Close();
 
 			// Sets the values to the values that were in the file
-			PerLevelData[0] = Data.Level1;
-			PerLevelData[1] = Data.Level2;
-			PerLevelData[2] = Data.Level3;
-			PerLevelData[3] = Data.Level4;
-			PerLevelData[4] = Data.Level5;
-			PerLevelData[5] = Data.Level6;
-			PerLevelData[6] = Data.Level7;
-			PerLevelData[7] = Data.Level8;
-			PerLevelData[8] = Data.Level9;
-			PerLevelData[9] = Data.Level10;
-			PerLevelData[10] = Data.Level11;
-			PerLevelData[11] = Data.Level12;
-			PerLevelData[12] = Data.Level13;
-			PerLevelData[13] = Data.Level14;
-			PerLevelData[14] = Data.Level15;
-			PerLevelData[15] = Data.Level16;
-			PerLevelData[16] = Data.Level17;
-			PerLevelData[17] = Data.Level18;
-			PerLevelData[18] = Data.Level19;
-			PerLevelData[19] = Data.Level20;
+			LevelData[0] = Data.Level1;
+			LevelData[1] = Data.Level2;
+			LevelData[2] = Data.Level3;
+			LevelData[3] = Data.Level4;
+			LevelData[4] = Data.Level5;
+			LevelData[5] = Data.Level6;
+			LevelData[6] = Data.Level7;
+			LevelData[7] = Data.Level8;
+			LevelData[8] = Data.Level9;
+			LevelData[9] = Data.Level10;
+			LevelData[10] = Data.Level11;
+			LevelData[11] = Data.Level12;
+			LevelData[12] = Data.Level13;
+			LevelData[13] = Data.Level14;
+			LevelData[14] = Data.Level15;
+			LevelData[15] = Data.Level16;
+			LevelData[16] = Data.Level17;
+			LevelData[17] = Data.Level18;
+			LevelData[18] = Data.Level19;
+			LevelData[19] = Data.Level20;
 		}
 	}
 }
 
-// holds things we wanna save :)
+// Class that stores the values 
 [Serializable]
 class GameData
 {
